move runAtStartup file into src
